refactor: console.log 정리 및 types 정리
